LayoutTests/http/tests/security/credentials-iframes-expected.txt

ALERT: parent host: 127.0.0.1 iframe host: 127.0.0.1 credentials:User: same-domain-user, password: same-domain-password.
127.0.0.1:8000 - didReceiveAuthenticationChallenge - Simulating cancelled authentication sheet
ALERT: parent host: localhost iframe host: 127.0.0.1 credentials:Authentication canceled
